Template: Add new scheduler docs
